requirements.txt
streamlit>=1.28.0
streamlit-extras>=0.3.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.3.0
PyPDF2>=3.0.0
chromadb>=0.4.0 
tavily-python>=0.1.0